Tightened cuts to fit triggers
no need for such loose cuts anymore

+++ Skimming/python/MUSiCSkimmer_cfi.py
@@ -60,16 +60,16 @@ Skimmer = cms.EDAnalyzer(
 
     cuts = cms.PSet(
         min_muon_pt = cms.double( 5 ),
-        min_ele_pt = cms.double( 5 ),
-        min_gamma_pt = cms.double( 5 ),
-        min_jet_pt = cms.double( 20 ),
-        min_met = cms.double( 20 ),
+        min_ele_pt = cms.double( 15 ),
+        min_gamma_pt = cms.double( 15 ),
+        min_jet_pt = cms.double( 80 ),
+        min_met = cms.double( 45 ),
         max_eta = cms.double( 3 ),
         vertex_minNDOF = cms.double( 3 ),
-        vertex_maxZ = cms.double( 20 ),
-        vertex_maxR = cms.double( 4 ),
+        vertex_maxZ = cms.double( 30 ),
+        vertex_maxR = cms.double( 3 ),
         PV_minNDOF = cms.double( 5 ),
-        PV_maxZ = cms.double( 15 ),
+        PV_maxZ = cms.double( 24 ),
         PV_maxR = cms.double( 2 )
         )
     )
